requirements.txt
asgiref==3.8.1
dj-config-url==0.1.1
dj-environ==0.0.1
Django==5.0.4
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
gunicorn==22.0.0
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
whitenoise==6.6.0
